requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_bcrypt==1.0.1
WTForms==2.3.3
Werkzeug==2.2.2